chore: Move StartUIIconCell into its own file
