+++ wire-ios/Wire-iOS/Sources/UserInterface/StartUI/Common/Cells/InviteTeamMemberCell.swift
@@ -19,77 +19,6 @@
 import UIKit
 import WireCommonComponents
 
-class StartUIIconCell: UICollectionViewCell {
-
-    typealias CellColors = SemanticColors.View
-    typealias PeoplePicker = L10n.Localizable.Peoplepicker
-
-    fileprivate let iconView = UIImageView()
-    fileprivate let titleLabel = DynamicFontLabel(fontSpec: .bodyTwoSemibold,
-                                                  color: SemanticColors.Label.textDefault)
-    fileprivate let separator = UIView()
-
-    fileprivate var icon: StyleKitIcon? {
-        didSet {
-            iconView.image = icon?.makeImage(size: .tiny, color: SemanticColors.Icon.foregroundDefault).withRenderingMode(.alwaysTemplate)
-            iconView.tintColor = SemanticColors.Icon.foregroundDefault
-        }
-    }
-
-    fileprivate var title: String? {
-        didSet {
-            titleLabel.text = title
-        }
-    }
-
-    override var isHighlighted: Bool {
-        didSet {
-            backgroundColor = isHighlighted ? CellColors.backgroundUserCellHightLighted : CellColors.backgroundUserCell
-        }
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupViews()
-        createConstraints()
-    }
-
-    @available(*, unavailable)
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    fileprivate func setupViews() {
-        iconView.contentMode = .center
-        separator.backgroundColor = CellColors.backgroundSeparatorCell
-        backgroundColor = CellColors.backgroundUserCell
-        [iconView, titleLabel, separator].forEach(contentView.addSubview)
-    }
-
-    fileprivate func createConstraints() {
-        let iconSize: CGFloat = 32.0
-
-        [iconView, titleLabel, separator].forEach { $0.translatesAutoresizingMaskIntoConstraints = false }
-        NSLayoutConstraint.activate([
-            iconView.widthAnchor.constraint(equalToConstant: iconSize),
-            iconView.heightAnchor.constraint(equalToConstant: iconSize),
-            iconView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-            iconView.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
-
-            titleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 64),
-            titleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
-            titleLabel.topAnchor.constraint(equalTo: contentView.topAnchor),
-            titleLabel.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
-
-            separator.leadingAnchor.constraint(equalTo: titleLabel.leadingAnchor),
-            separator.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
-            separator.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
-            separator.heightAnchor.constraint(equalToConstant: .hairline)
-        ])
-    }
-
-}
-
 final class InviteTeamMemberCell: StartUIIconCell {
 
     override func setupViews() {
